forms: Fix fetching To: and Cc: from a form
There are forms that have "Send To" or whatnot in them that are supposed
to fill in the To: field of the message, but this wasn't working.
There were two issues:

The token field config name for delimiters is "delimiters", not
"delimiter".  That lets the field get set, but only ';' would work as a
delimiter.

Modify the split operation for msg_to and msg_cc to use a regular
expression with ' ', ';', and ',' per how the rest of the system works.

Signed-off-by: Corey Minyard <cminyard@mvista.com>